rename vst_api to proof_plumber_api
