+++ crates/ide-assists/src/assist_context.rs
@@ -9,7 +9,7 @@ use syntax::{
     TextSize, TokenAtOffset,
 };
 
-use crate::vst_api::verus_error::VerusError;
+use crate::proof_plumber_api::verus_error::VerusError;
 use crate::{
     assist_config::AssistConfig, Assist, AssistId, AssistKind, AssistResolveStrategy, GroupLabel,
 };
