+++ crates/ide-assists/src/handlers/proof_action/weakest_pre_step.rs
@@ -1,6 +1,6 @@
 use crate::{AssistContext, Assists};
 use ide_db::assists::{AssistId, AssistKind};
-use crate::vst_api::vst_ext::*;
+use crate::proof_plumber_api::vst_ext::*;
 use syntax::{ast::{self, vst::*}, AstNode, T};
 
 /*
