+++ crates/ide-assists/src/proof_plumber_api/inline_function_api.rs
