+++ crates/ide-assists/src/proof_plumber_api/semantic_info.rs
