+++ crates/ide-assists/src/proof_plumber_api/vst_ext.rs
