+++ crates/ide-assists/src/tests.rs
@@ -17,7 +17,7 @@ use syntax::TextRange;
 use test_utils::{assert_eq_text, extract_offset};
 
 use crate::{
-    assists, handlers::Handler, vst_api::verus_error::VerusError, Assist, AssistConfig, AssistContext, AssistKind, AssistResolveStrategy, Assists, SingleResolve
+    assists, handlers::Handler, proof_plumber_api::verus_error::VerusError, Assist, AssistConfig, AssistContext, AssistKind, AssistResolveStrategy, Assists, SingleResolve
 };
