+++ crates/ide/src/lib.rs
@@ -65,7 +65,7 @@ use std::ffi::OsStr;
 
 use cfg::CfgOptions;
 use fetch_crates::CrateInfo;
-use ide_assists::vst_api::verus_error::VerusError;
+use ide_assists::proof_plumber_api::verus_error::VerusError;
 use ide_db::{
     base_db::{
         salsa::{self, ParallelDatabase},
