+++ crates/rust-analyzer/src/global_state.rs
@@ -120,7 +120,7 @@ pub(crate) struct GlobalState {
     pub(crate) prime_caches_queue: OpQueue,
 
     // verus
-    pub(crate) verus_errors: Vec<ide_assists::vst_api::verus_error::VerusError>,
+    pub(crate) verus_errors: Vec<ide_assists::proof_plumber_api::verus_error::VerusError>,
 }
 
 /// An immutable snapshot of the world's state at a point in time.
@@ -136,7 +136,7 @@ pub(crate) struct GlobalStateSnapshot {
     pub(crate) proc_macros_loaded: bool,
     pub(crate) flycheck: Arc<[FlycheckHandle]>,
     // verus
-    pub(crate) verus_errors: Vec<ide_assists::vst_api::verus_error::VerusError>,
+    pub(crate) verus_errors: Vec<ide_assists::proof_plumber_api::verus_error::VerusError>,
 }
 
 impl std::panic::UnwindSafe for GlobalStateSnapshot {}
